clinicBot/requirements.txt
flask 
requests 
langchain 
faiss-cpu 
openai 
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client 
python-dotenv
